refactor: util함수 합치기
