--- src/utils/accountActive.ts
@@ -1,3 +1,0 @@
-export default function accountActive(isActive: boolean) {
-  return isActive ? '활성화' : '비활성화';
-}
