--- src/utils/comma.ts
@@ -1,3 +1,0 @@
-export default function comma(price: string) {
-  return parseInt(price)?.toLocaleString();
-}
