--- src/utils/dateFormat.ts
@@ -1,3 +1,0 @@
-export default function dateFormat(date: string) {
-  return date?.slice(0, 10);
-}
